refactor(observe): extract shared filter state into useObserveFilters
InsightsTools and LogsTools duplicated ~300 lines of identical filter
logic (URL param parsing, date range, server/email/type filters,
knownServers accumulation)
