+++ client/dashboard/src/components/observe/InsightsTools.tsx
@@ -26,7 +26,6 @@ import type {
   SkillBreakdownRow,
   SkillTimeSeriesPoint,
   HookTraceSummary as HookTrace,
-  LogFilter,
   TelemetryLogRecord,
   TypesToInclude,
 } from "@gram/client/models/components";
@@ -56,7 +55,9 @@ import {
 import { Bar, Line } from "react-chartjs-2";
 import { Settings } from "lucide-react";
 import { useCallback, useEffect, useMemo, useState } from "react";
-import { Link, useSearchParams } from "react-router";
+import { Link } from "react-router";
+import { useObserveFilters } from "@/components/observe/useObserveFilters";
+import { perPage } from "@/components/observe/observeFilterUtils";
 import { LogDetailSheet } from "@/pages/logs/LogDetailSheet";
 import { HooksEmptyState } from "@/pages/hooks/HooksEmptyState";
 import { HooksSetupButton } from "@/pages/hooks/HooksSetupDialog";
@@ -163,48 +164,7 @@ const SHARED_BAR_SCALES = {
   },
 } satisfies _BarScales;
 
-const validPresets: DateRangePreset[] = [
-  "15m",
-  "1h",
-  "4h",
-  "1d",
-  "2d",
-  "3d",
-  "7d",
-  "15d",
-  "30d",
-  "90d",
-];
-
-function isValidPreset(value: string | null): value is DateRangePreset {
-  return value !== null && validPresets.includes(value as DateRangePreset);
-}
-
-function safeBase64Encode(str: string): string {
-  try {
-    return btoa(str);
-  } catch {
-    return btoa(encodeURIComponent(str));
-  }
-}
-
-function safeBase64Decode(str: string): string | null {
-  try {
-    const decoded = atob(str);
-    try {
-      return decodeURIComponent(decoded);
-    } catch {
-      return decoded;
-    }
-  } catch {
-    return null;
-  }
-}
-
-const perPage = 100;
-
 export function InsightsToolsContent() {
-  const [searchParams, setSearchParams] = useSearchParams();
   const { projectSlug } = useSlugs();
 
   const mcpConfig = useObservabilityMcpConfig({
@@ -214,140 +174,34 @@ export function InsightsToolsContent() {
 
   const serverNameMappings = useServerNameMappings();
 
-  const initialServer = searchParams.get("server");
-  const initialUserEmail = searchParams.get("user");
+  const {
+    from,
+    to,
+    logFilters,
+    selectedHookTypes,
+    activeFilters,
+    addKnownServers,
+    serverOptions,
+    handleServerSelectionChange,
+    userEmailInput,
+    setUserEmailInput,
+    submitUserEmailFilter,
+    addFilter,
+    removeFilter,
+    handleHookTypesChange,
+    dateRange,
+    customRange,
+    customRangeLabel,
+    setDateRangeParam,
+    setCustomRangeParam,
+    clearCustomRange,
+  } = useObserveFilters();
 
-  const initialHookTypes = searchParams.get("hookTypes");
-  const defaultHookTypes: TypesToInclude[] = ["mcp", "skill"];
-  const parsedHookTypes: TypesToInclude[] = initialHookTypes
-    ? (initialHookTypes
-        .split(",")
-        .filter((t) =>
-          ["mcp", "local", "skill"].includes(t),
-        ) as TypesToInclude[])
-    : defaultHookTypes;
-
-  const [activeFilters, setActiveFilters] = useState<FilterChip[]>(() => {
-    const filters: FilterChip[] = [];
-
-    if (initialServer) {
-      const serverValues = initialServer
-        .split(",")
-        .map((s) => s.trim())
-        .filter(Boolean);
-      serverValues.forEach((value) => {
-        filters.push({
-          display: value,
-          filters: [value],
-          path: "gram.tool_call.source",
-        });
-      });
-    }
-
-    if (initialUserEmail) {
-      const userValues = initialUserEmail
-        .split(",")
-        .map((s) => s.trim())
-        .filter(Boolean);
-      userValues.forEach((value) => {
-        filters.push({
-          display: value,
-          filters: [value],
-          path: "user.email",
-        });
-      });
-    }
-
-    return filters;
-  });
-
-  const [userEmailInput, setUserEmailInput] = useState("");
-  const [selectedHookTypes, setSelectedHookTypes] =
-    useState<TypesToInclude[]>(parsedHookTypes);
   const [selectedLog, setSelectedLog] = useState<TelemetryLogRecord | null>(
     null,
   );
 
   const client = useGramContext();
-
-  const urlRange = searchParams.get("range");
-  const urlFrom = searchParams.get("from");
-  const urlTo = searchParams.get("to");
-  const urlLabelEncoded = searchParams.get("label");
-  const urlLabel = useMemo(() => {
-    if (!urlLabelEncoded) return null;
-    return safeBase64Decode(urlLabelEncoded);
-  }, [urlLabelEncoded]);
-
-  const dateRange: DateRangePreset = isValidPreset(urlRange) ? urlRange : "7d";
-
-  const customRange = useMemo(() => {
-    if (urlFrom && urlTo) {
-      const from = new Date(urlFrom);
-      const to = new Date(urlTo);
-      if (!isNaN(from.getTime()) && !isNaN(to.getTime())) {
-        return { from, to };
-      }
-    }
-    return null;
-  }, [urlFrom, urlTo]);
-
-  const updateSearchParams = useCallback(
-    (updates: Record<string, string | null>) => {
-      setSearchParams((prev) => {
-        const next = new URLSearchParams(prev);
-        for (const [key, value] of Object.entries(updates)) {
-          if (value === null) {
-            next.delete(key);
-          } else {
-            next.set(key, value);
-          }
-        }
-        return next;
-      });
-    },
-    [setSearchParams],
-  );
-
-  const setDateRangeParam = useCallback(
-    (preset: DateRangePreset) => {
-      updateSearchParams({ range: preset, from: null, to: null, label: null });
-    },
-    [updateSearchParams],
-  );
-
-  const setCustomRangeParam = useCallback(
-    (from: Date, to: Date, label?: string) => {
-      updateSearchParams({
-        range: null,
-        from: from.toISOString(),
-        to: to.toISOString(),
-        label: label ? safeBase64Encode(label) : null,
-      });
-    },
-    [updateSearchParams],
-  );
-
-  const clearCustomRange = useCallback(() => {
-    updateSearchParams({ from: null, to: null, label: null });
-  }, [updateSearchParams]);
-
-  const { from, to } = useMemo(
-    () => customRange ?? getPresetRange(dateRange),
-    [customRange, dateRange],
-  );
-
-  const logFilters = useMemo(() => {
-    const filters: LogFilter[] = [];
-    for (const chip of activeFilters) {
-      filters.push({
-        path: chip.path,
-        operator: chip.filters.length > 1 ? "in" : "contains",
-        values: chip.filters,
-      });
-    }
-    return filters.length > 0 ? filters : undefined;
-  }, [activeFilters]);
 
   const {
     data: tracesData,
@@ -416,165 +270,10 @@ export function InsightsToolsContent() {
     throwOnError: false,
   });
 
-  const [knownServers, setKnownServers] = useState<string[]>([]);
-
   useEffect(() => {
     if (!summaryData) return;
-    const names = summaryData.breakdown.map((r) => r.serverName);
-    setKnownServers((prev) => {
-      const merged = [...new Set([...prev, ...names])];
-      return merged.length === prev.length ? prev : merged;
-    });
-  }, [summaryData]);
-
-  const serverOptions = useMemo(() => {
-    const selected = activeFilters
-      .filter((f) => f.path === "gram.tool_call.source")
-      .map((f) => f.filters[0])
-      .filter((v): v is string => Boolean(v));
-    return [...new Set([...knownServers, ...selected])];
-  }, [knownServers, activeFilters]);
-
-  const handleServerSelectionChange = useCallback(
-    (values: string[]) => {
-      setActiveFilters((prev) => {
-        const nonServer = prev.filter(
-          (f) => f.path !== "gram.tool_call.source",
-        );
-        const serverFilters: FilterChip[] = values.map((v) => ({
-          display: v,
-          filters: [v],
-          path: "gram.tool_call.source",
-        }));
-        return [...nonServer, ...serverFilters];
-      });
-      setSearchParams(
-        (urlPrev) => {
-          const next = new URLSearchParams(urlPrev);
-          if (values.length > 0) {
-            next.set("server", values.join(","));
-          } else {
-            next.delete("server");
-          }
-          return next;
-        },
-        { replace: true },
-      );
-    },
-    [setSearchParams],
-  );
-
-  const addFilter = useCallback(
-    (chip: FilterChip) => {
-      setActiveFilters((prev) => {
-        const exists = prev.some(
-          (f) => f.path === chip.path && f.display === chip.display,
-        );
-        if (exists) return prev;
-
-        const newFilters = [...prev, chip];
-
-        setSearchParams(
-          (urlPrev) => {
-            const next = new URLSearchParams(urlPrev);
-            if (chip.path === "gram.tool_call.source") {
-              const serverFilters = newFilters
-                .filter((f) => f.path === "gram.tool_call.source")
-                .map((f) => f.display);
-              next.set("server", serverFilters.join(","));
-            } else if (chip.path === "user.email") {
-              const userFilters = newFilters
-                .filter((f) => f.path === "user.email")
-                .map((f) => f.display);
-              next.set("user", userFilters.join(","));
-            }
-            return next;
-          },
-          { replace: true },
-        );
-
-        return newFilters;
-      });
-    },
-    [setSearchParams],
-  );
-
-  const removeFilter = useCallback(
-    (path: string, display?: string) => {
-      setActiveFilters((prev) => {
-        const newFilters = display
-          ? prev.filter((f) => !(f.path === path && f.display === display))
-          : prev.filter((f) => f.path !== path);
-
-        setSearchParams(
-          (urlPrev) => {
-            const next = new URLSearchParams(urlPrev);
-            if (path === "gram.tool_call.source") {
-              const serverFilters = newFilters
-                .filter((f) => f.path === "gram.tool_call.source")
-                .map((f) => f.display);
-              if (serverFilters.length > 0) {
-                next.set("server", serverFilters.join(","));
-              } else {
-                next.delete("server");
-              }
-            } else if (path === "user.email") {
-              const userFilters = newFilters
-                .filter((f) => f.path === "user.email")
-                .map((f) => f.display);
-              if (userFilters.length > 0) {
-                next.set("user", userFilters.join(","));
-              } else {
-                next.delete("user");
-              }
-            }
-            return next;
-          },
-          { replace: true },
-        );
-
-        return newFilters;
-      });
-    },
-    [setSearchParams],
-  );
-
-  const submitUserEmailFilter = useCallback(() => {
-    const trimmed = userEmailInput.trim();
-    if (!trimmed) return;
-    addFilter({
-      display: trimmed,
-      filters: [trimmed],
-      path: "user.email",
-    });
-    setUserEmailInput("");
-  }, [userEmailInput, addFilter]);
-
-  const handleHookTypesChange = useCallback(
-    (types: TypesToInclude[]) => {
-      setSelectedHookTypes(types);
-      setSearchParams(
-        (prev) => {
-          const next = new URLSearchParams(prev);
-          const isDefault =
-            types.length === 2 &&
-            types.includes("mcp") &&
-            types.includes("skill") &&
-            !types.includes("local");
-          if (isDefault) {
-            next.delete("hookTypes");
-          } else if (types.length > 0) {
-            next.set("hookTypes", types.join(","));
-          } else {
-            next.set("hookTypes", "");
-          }
-          return next;
-        },
-        { replace: true },
-      );
-    },
-    [setSearchParams],
-  );
+    addKnownServers(summaryData.breakdown.map((r) => r.serverName));
+  }, [summaryData, addKnownServers]);
 
   const refetch = useCallback(() => {
     refetchLogs();
@@ -633,7 +332,7 @@ export function InsightsToolsContent() {
             setSelectedLog={setSelectedLog}
             dateRange={dateRange}
             customRange={customRange}
-            customRangeLabel={urlLabel}
+            customRangeLabel={customRangeLabel}
             onDateRangeChange={setDateRangeParam}
             onCustomRangeChange={setCustomRangeParam}
             onClearCustomRange={clearCustomRange}
